Update referral-bonuses.md
Updated the new hire referral language to be a bit more clear. Increased qualifying hire date from 45 days to 60 days, in consideration of the fact that our time-to-hire is fairly lengthy.

Scrapped old new client referral policy and replaced it with the same referral partner program we offer to our most loyal/trusted clients. The new program is much more generous and potentially lucrative.